Rename source files to .cpp to avoid needing to force C++ compiler on the command line.
