+++ source/DemodFSK.cpp
@@ -38,7 +38,7 @@
 */
 #include <math.h>
 #include <mex.h>
-#include <matrix.h>
+#include <Matrix.h>
 #include <stdlib.h>
 
 /* Input Arguments
