+++ source/LdpcEncode.cpp
