+++ source/MpDecode.cpp
@@ -51,7 +51,7 @@
 
 #include <math.h>
 #include <mex.h>
-#include <matrix.h>
+#include <Matrix.h>
 #include <stdlib.h>
 #include <stdio.h>
