+++ source/ViterbiDecode.cpp
@@ -44,7 +44,7 @@
 
 #include <math.h>
 #include <mex.h>
-#include <matrix.h>
+#include <Matrix.h>
 #include <stdlib.h>
 
 /* library of functions */
